Share selectConsolidatedVersions with the backend
Move selectConsolidatedVersions/todayIso out of the frontend-only
src/utils/consolidatedVersions.js into backend/shared/consolidated-versions.mjs
(ESM, named exports, no deps), and re-export from the frontend file —
mirroring src/utils/url.js re-exporting formex-parser/url.mjs. Prep for the
next commit, which needs the same "pick the current, non-future-dated
version" logic on the backend to choose a consolidated fallback.

summarizeAmendments stays frontend-only (nothing backend needs it), and
consolidatedVersions.test.js keeps importing the frontend path unchanged.
