+++ src/utils/consolidatedVersions.js
@@ -7,40 +7,13 @@
  * one) is still readable as adopted. *Can the reader see the current text?*
  * comes from the consolidated-version list.
  *
- * The API returns every consolidated version, including ones dated in the
- * future — EUR-Lex prepares a consolidation as soon as an amending act is
- * published, sometimes months before it applies. Choosing between them needs
- * today's date, which deliberately stays out of the cached payload.
+ * `todayIso` and `selectConsolidatedVersions` live in
+ * `backend/shared/consolidated-versions.mjs` (shared with the backend
+ * consolidated-fallback path in `parsed-law-service.js`) and are re-exported
+ * here, mirroring `src/utils/url.js` re-exporting `formex-parser/url.mjs`.
  */
 
-/** ISO `YYYY-MM-DD` for the local calendar day, so comparisons are string-safe. */
-export function todayIso(now = new Date()) {
-  const pad = (value) => String(value).padStart(2, "0");
-  return `${now.getFullYear()}-${pad(now.getMonth() + 1)}-${pad(now.getDate())}`;
-}
-
-/**
- * Split a version list into the one in force today and the ones not yet applied.
- *
- * `current` is the newest version dated on or before today. A law whose only
- * consolidations are future-dated has no current version — presenting an
- * upcoming one as the text in force would be worse than saying nothing.
- */
-export function selectConsolidatedVersions(versions, today = todayIso()) {
-  const sorted = (Array.isArray(versions) ? versions : [])
-    .filter((version) => version && version.celex && version.date)
-    .slice()
-    .sort((a, b) => a.date.localeCompare(b.date));
-
-  const applied = sorted.filter((version) => version.date <= today);
-  const upcoming = sorted.filter((version) => version.date > today);
-
-  return {
-    current: applied.length ? applied[applied.length - 1] : null,
-    upcoming,
-    all: sorted,
-  };
-}
+export * from "../../backend/shared/consolidated-versions.mjs";
 
 /**
  * Count the entries that actually changed the law's text, and date the newest.
